isolate demo prefabs, editor and test runner into classic-demo
Move the demo/application layer out of `classic-engine` so the
engine is demo-neutral and `classic-demo` owns all prefab,
editor, lighting, overlay and test-scenario code.

- `Engine` drops its demo state (`editor`, widget handles, compass
  buffer, test-harness fields) and gains a generic hook surface:
  `on_pre_update`, `on_selection_end`, `add_overlay`,
  `set_test_runner`, plus `agent_selected` and `frame_number` /
  `viewport_size` accessors.
- `frame()` now dispatches those hooks instead of inlining the
  text-scroll routing, editor paint, footprint/agent-ring/compass
  overlays and the CLASSIC_TEST block.
- `classic-demo` gains `state` (DemoState + EditorState), `prefabs`,
  `lighting`, `editor`, `hud` and `testing` modules, written as
  free functions over `&mut Engine` + `Rc<RefCell<DemoState>>`.
- `init_engine` composes them and registers the hooks; the nav
  click-to-move now reads `engine.agent_selected` instead of the
  editor field.

Pure relocation: no behaviour change, demo golden trace is
byte-identical (12/12 assertions + trace match).

(this patch was generated in some part by `opencode` using `deepseek-v4-pro` (`deepseek`))
