+++ crates/classic-demo/Cargo.toml
@@ -13,3 +13,5 @@ glow.workspace = true
 hecs.workspace = true
 anyhow.workspace = true
 log.workspace = true
+serde.workspace = true
+serde_json.workspace = true
