--- crates/classic-engine/src/testing/mod.rs
@@ -1,7 +1,0 @@
-//! # Skill: `classic-testing`
-//!
-//! **Read `.claude/skills/classic-testing/SKILL.md` before working on this module.**
-//!
-//! CLASSIC_TEST framework: types, scenario builder, assertion runner.
-
-pub mod types;
